Don't use JsNumber

+++ facade/src/main/scala/react/gridlayout/ResponsiveReactGridLayout.scala
@@ -12,7 +12,7 @@ import react.common._
 import japgolly.scalajs.react.vdom.TagMod
 
 final case class ResponsiveReactGridLayout(
-  width:                  Int,
+  width:                  Double,
   layouts:                Map[BreakpointName, (Int, Int, Layout)],
   className:              js.UndefOr[String] = js.undefined,
   style:                  js.UndefOr[Style] = js.undefined,
@@ -123,7 +123,7 @@ object ResponsiveReactGridLayout {
     )
 
   def rawprops(
-    width:              Int,
+    width:              Double,
     layouts:            Map[BreakpointName, (Int, Int, Layout)],
     className:          js.UndefOr[String] = js.undefined,
     style:              js.UndefOr[Style] = js.undefined,
@@ -228,7 +228,7 @@ object ResponsiveReactGridLayout {
   val component = JsComponent[ResponsiveReactGridLayoutProps, Children.Varargs, Null](RawComponent)
 
   def apply(
-    width:   Int,
+    width:   Double,
     layouts: Map[BreakpointName, (Int, Int, Layout)],
     content: TagMod*
   ): ResponsiveReactGridLayout =
